A media type and a parameter are two questions, so two modules
The plan had these as one cluster. They are not, and the caller counts say so
in the wrong direction at first glance: `parameters` appears in far more files
than `parse_media_type`, and all but one of them parse a media type too. That
reads like an argument for a single `media_type` module.

It is an argument about which caller is busiest, not about what the function
asks. `parameters(params: &str)` takes any `;`-separated tail and knows nothing
of what preceded it — it is RFC 9110 § 5.6.6, a production `Content-Disposition`
and `Expect` also carry. Filing it under `media_type` would have named it for
its dominant reader and left `expect_header_valid` importing a media type to
read an `Expect` parameter.

So `helpers::parameter` holds the production and its typed `ParameterDefect`,
and `helpers::media_type` holds what makes a media type one — the slash, the
two halves, the `+suffix`, the multipart `boundary` — and depends on it. The
dependency now points the way the grammar does.

Cohesion check, again unprompted by any tool: after the move `headers.rs` has
no remaining caller of `parameters` or `ParameterDefect` at all.

headers.rs: 2344 → 1410 non-test lines. Crate total still 6032 tests.
